(fix) Fixed getPosts params to specify boardId pointer
Using params allows us to send data in a GET request, as Angular will encode the data into the url. The boardId is actually a reference to a Pointer object in Parse, so passing in a string does not work. You have to pass an object with the specified fields per the Parse REST API docs.